NCBC-2921: Value was either too large or too small for an Int32
Motivation
----------
Fixes a bug where the quota.rawRAM size may over/under flow the Int32
size of the BucketSettings.RamQuotaMB field when the JSON is parsed.

Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/157288
Tested-by: Build Bot <build@couchbase.com>

+++ src/Couchbase/Management/Buckets/BucketManager.cs
@@ -53,7 +53,7 @@ namespace Couchbase.Management.Buckets
             {
                 Name = json.GetTokenValue<string>("name"),
                 MaxTtl = json.GetTokenValue<int>("maxTTL"),
-                RamQuotaMB = json.GetTokenValue<int>("quota.rawRAM"),
+                RamQuotaMB = json.GetTokenValue<long>("quota.rawRAM"),
                 FlushEnabled = json.SelectToken("controllers.flush") != null
             };
